requirements.txt
fastapi
uvicorn
openai
paddleocr
paddlepaddle
opencv-python-headless
python-multipart
python-dotenv
requests
httpx
